Pre RC2 Update

+++ SMS/src/SMS/Controllers/HomeController.cs
@@ -1,4 +1,5 @@
-﻿using System.Linq;
+﻿using System;
+using System.Linq;
 using Microsoft.AspNet.Mvc;
 using SMS.Models;
 
@@ -42,47 +43,51 @@ namespace SMS.Controllers
             return View();
         }
 
+
         public IActionResult Seed([FromRoute] bool recreate = false)
         {
-             _db.Database.EnsureDeleted();
-                _db.Database.EnsureCreated();
-            var w1 = new Waypoint {Description = "Desc", Latitude = 50, Longitude = 11, Name = "W"};
-            var h1 = new Harbour
+            if (Startup.IsDevelopment)
             {
-                Description = "Desc",
-                Latitude = 50,
-                Longitude = 10,
-                Name = "H1",
-                Locations =
+                _db.Database.EnsureDeleted();
+                _db.Database.EnsureCreated();
+                var w1 = new Waypoint { Description = "Desc", Latitude = 50, Longitude = 11, Name = "W" };
+                var h1 = new Harbour
                 {
-                    new Location()
+                    Description = "Desc",
+                    Latitude = 50,
+                    Longitude = 10,
+                    Name = "H1",
+                    Locations =
                     {
-                        Name = "N",
-                        Type = LocationType.Restaurant,
-                        Address = new Address
+                        new Location()
                         {
-                            Comment = "C",
-                            Street = "S",
-                            Town = "T",
-                            Zip = "Z"
+                            Name = "N",
+                            Type = LocationType.Restaurant,
+                            Address = new Address
+                            {
+                                Comment = "C",
+                                Street = "S",
+                                Town = "T",
+                                Zip = "Z"
+                            }
                         }
                     }
-                }
-            };
-            var h2 = new Harbour
-            {
-                Description = "Desc",
-                Latitude = 50,
-                Longitude = 12,
-                Name = "H2",
-                Album = new Album()
-            };
-            _db.AddHarbour(h1);
-            _db.AddHarbour(h2);
-            _db.AddWaypoint(w1);
-            _db.WaypointConnections.Add(new WaypointConnection {Waypoint1 = h1, Waypoint2 = w1});
-            _db.WaypointConnections.Add(new WaypointConnection {Waypoint1 = h2, Waypoint2 = w1});
-            _db.SaveChanges();
+                };
+                var h2 = new Harbour
+                {
+                    Description = "Desc",
+                    Latitude = 50,
+                    Longitude = 12,
+                    Name = "H2",
+                    Album = new Album()
+                };
+                _db.AddHarbour(h1);
+                _db.AddHarbour(h2);
+                _db.AddWaypoint(w1);
+                _db.WaypointConnections.Add(new WaypointConnection { Waypoint1 = h1, Waypoint2 = w1 });
+                _db.WaypointConnections.Add(new WaypointConnection { Waypoint1 = h2, Waypoint2 = w1 });
+                _db.SaveChanges();
+            }
             return View("Index");
         }
     }
